Moved achievement-get functionality into component
Honestly not much more I can say here, the title says it all,
