+++ src/pages/home.svelte
@@ -20,15 +20,15 @@
             <Profile bind:this={profile}></Profile>
         </div>
       </section>
-
+      <AchievementGet></AchievementGet>
       <!-- Achievement-Get -->
-      <Modal bind:this={modal}>
+      <!-- <Modal bind:this={modal}>
           <div class="achievement-get">
               <h1 class="badge-notif">Badge Unlocked!</h1>
               <img id="badge-img" src="../static/badge/sample-badge.svg" alt="badge">
               <h3 class="badge-name">Badge Name Here</h3>
           </div>
-      </Modal>
+      </Modal> -->
 
       <!-- Current-Location and Bottom-Card -->
       <section class="bottom-card">
@@ -108,7 +108,7 @@
     font-weight: 500;
   }
 /* to be exported to achievement-get */
-  .achievement-get {
+  /* .achievement-get {
       display: flex;
       flex-direction: column;
   }
@@ -126,7 +126,7 @@
   }
   #badge-img {
       height: 60%;
-  }
+  } */
 
       /* Sections */
 
@@ -170,9 +170,8 @@
   // importing functionality
   import Landing from '../components/landing.svelte';
   import LoadingIcon from '../components/loading.svelte';
-  // import AchievementGet from '../components/achievement-get.svelte'; needs to be created
+  import AchievementGet from '../components/modal/achievement-get.svelte';
   import Profile from '../components/profile.svelte';
-
   import ListBadges from '../components/listBadges.svelte';
 
   let profile;
@@ -182,11 +181,11 @@
   });
 
   // move to achievement-get component
-  import Modal from "../components/modal.svelte";
-  let modal;
+  // import Modal from "../components/modal.svelte";
+  // let modal;
   
-  let badgeNotificationAudio = new Audio("../static/audio/app_alert_tone_011.mp3");
-  function playBadgeNotificationAudio() {
-    badgeNotificationAudio.play();
-  }
+  // let badgeNotificationAudio = new Audio("../static/audio/app_alert_tone_011.mp3");
+  // function playBadgeNotificationAudio() {
+  //   badgeNotificationAudio.play();
+  // }
 </script>
